requirements.txt
Flask
matplotlib
mysql-connector-python